Add tasks to install adafruit hostapd
All my local USB wifi dongles are realtek-based. This chipset
is not supported by the stock hostapd in raspbian. Therefore
I fetch, unpack and install the adafruit version of hostapd that
contains realtek support.

This is the most complex type of installation and must be optional
in the long run.