Add script to map nic names on macs.
We also show drivers and buses used by respective network devices.

This info is needed in order to determine the built-in wifi device
in a raspi. If you use multiple wifi dongles at the same time, the
kernel will name these "wlan0", "wlan1", etc. But there is no
guarantee, that "wlan0" is the built-in wifi. And even if it is,
after a reboot things could change completely.

With the infos from this script we can pin down certain devices
to certain device names and be sure, that it is always the same
device.