Add task to pin wlan0 to builtin wifi device.
We use multiple wifi devices and must make sure that the builtin device
is mapped to "wlan0". This is neccessary, as we must rely on interface
names in following steps.

If we do not pin the names, then the kernel might assign different names
to the same device on different runs.