Use original `hostapd`.
Do not install the patched Adafruit version but go with the Raspbian
original version of `hostapd`. The patched version contains non-free
driver code from Realtek. This is required if we want non-free USB wifi
adapters to act as as accesspoint (WLAN1). But we can also use the
builtin interface for this.