Replace old dynamic udev rule with static one.
The udev rule created by `install_udev_wlan0.sh` not only is named
erranously (because it creates a rule to determine `wlan1`, not
`wlan0`), it is also not neccessary at all.

The udev rule itself can check attributes of devices and does not need a
mac address if there are other attributes we can use. In fact the script
exemines attributes (drivername and bus), wich can be determined by the
udev rule itself.

As a side effect, we get a udev rule which can be used when building
Raspbian images.